+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `packagist-api` will be documented in this file
 
+## 2.0.0 - (unreleased)
+- Renamed `Packagist` class to `PackagistClient`.
+- Changed the interface of the `PackagistClient` class to reflect the current state of the [Packagist API](https://packagist.org/apidoc)
+- Moved url generation to a separate class.
+- Moved vendor formatting to a separate class.
+- Changed the meta data method to actually use the repository endpoint.
+- Split up integration and unit tests
+
 ## 1.3.1 - 2020-03-25
 
 - add support for PHP 7.2